Added parameter in GoToLevel to go just above that level.
Added ability for go to level to go slightly above a given level by adding an offset in the ClawPID subsystem to that levels height value. Giving go to level a "true" when called will cause it to go the offset distance above the specified level, "false" will result in the exact level value. Currently GToLevel will not add an offset to level 6 (the highest level, in order to prevent going outside the range of the elevator).